requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
docstring-parser==0.16
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-functions==1.20.3
google-cloud-resource-manager==1.14.2
google-cloud-scheduler==2.16.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.6
packaging==25.0
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1

# 새로 추가되는 라이브러리들
# FastAPI 및 웹 프레임워크
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0

# 소셜미디어 API 클라이언트
facebook-sdk==3.1.0
google-api-python-client==2.156.0
tweepy==4.14.0

# 데이터 처리 및 분석
scikit-learn==1.6.0
nltk==3.9.1
wordcloud==1.9.4
textblob==0.18.0

# 이메일 및 알림
sendgrid==6.12.1
jinja2==3.1.5

# 개발 도구
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.12.0
pylint==3.3.2
mypy==1.14.1

# 기타 유틸리티
python-dotenv==1.0.1
pydantic-settings==2.8.0
redis==5.2.1
celery==5.4.0
